# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  weightKg?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileInput {
  /** @minLength 1 */
  name: string;
  email: string;
  age?: number;
  gender?: string;
  heightCm?: number;
  weightKg?: number;
}

export interface Activity {
  id: number;
  /** @nullable */
  steps?: number | null;
  /** @nullable */
  distanceKm?: number | null;
  /** @nullable */
  caloriesBurned?: number | null;
  activityDate: string;
  createdAt: string;
  updatedAt: string;
}

export interface ActivityInput {
  steps?: number;
  distanceKm?: number;
  caloriesBurned?: number;
  activityDate: string;
}

export interface ActivityUpdate {
  steps?: number;
  distanceKm?: number;
  caloriesBurned?: number;
  activityDate?: string;
}

export interface Workout {
  id: number;
  workoutName: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  caloriesBurned?: number | null;
  workoutDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface WorkoutInput {
  /** @minLength 1 */
  workoutName: string;
  durationMinutes?: number;
  caloriesBurned?: number;
  workoutDate: string;
  notes?: string;
}

export interface WorkoutUpdate {
  /** @minLength 1 */
  workoutName?: string;
  durationMinutes?: number;
  caloriesBurned?: number;
  workoutDate?: string;
  notes?: string;
}

export interface Goals {
  id: number;
  /** @nullable */
  stepGoal?: number | null;
  /** @nullable */
  calorieGoal?: number | null;
  /** @nullable */
  waterGoalMl?: number | null;
  /** @nullable */
  workoutGoalPerWeek?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface GoalsInput {
  stepGoal?: number;
  calorieGoal?: number;
  waterGoalMl?: number;
  workoutGoalPerWeek?: number;
}

export interface WaterIntake {
  id: number;
  amountMl: number;
  loggedAt: string;
  createdAt: string;
}

export interface WaterIntakeInput {
  amountMl: number;
}

export interface WaterDaySummary {
  date: string;
  totalMl: number;
}

export interface DashboardSummary {
  todaySteps: number;
  todayCalories: number;
  todayDistanceKm: number;
  todayWaterMl: number;
  weeklyWorkouts: number;
  stepGoalPercent: number;
  calorieGoalPercent: number;
  waterGoalPercent: number;
  recentWorkouts: Workout[];
  recentActivities: Activity[];
}

export interface WeeklyReportDay {
  date: string;
  steps: number;
  calories: number;
  distanceKm: number;
  workoutCount: number;
}

export interface WeeklyReport {
  weekStart: string;
  weekEnd: string;
  totalSteps: number;
  totalCalories: number;
  totalDistanceKm: number;
  totalWorkouts: number;
  totalWaterMl: number;
  days: WeeklyReportDay[];
}

export interface MonthlyReportWeek {
  weekLabel: string;
  steps: number;
  calories: number;
  workouts: number;
}

export interface MonthlyReport {
  month: number;
  year: number;
  totalSteps: number;
  totalCalories: number;
  totalDistanceKm: number;
  totalWorkouts: number;
  totalWaterMl: number;
  weeks: MonthlyReportWeek[];
}

export interface BmiResult {
  bmi: number;
  category: string;
  heightCm: number;
  weightKg: number;
}

export interface StreakResult {
  /** Consecutive days the step goal has been hit (ending today or yesterday) */
  currentStreak: number;
  /** All-time longest streak */
  longestStreak: number;
  /** The step goal used to compute the streak (0 if no goal set) */
  stepGoal: number;
  /** Whether today's steps already meet the goal */
  goalHitToday: boolean;
}

export type ListActivitiesParams = {
limit?: number;
offset?: number;
};

export type ListWorkoutsParams = {
limit?: number;
offset?: number;
};

export type ListWaterIntakeParams = {
limit?: number;
};