# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityInput {
  steps?: number;
  distanceKm?: number;
  caloriesBurned?: number;
  activityDate: Date;
}