# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { Workout } from './workout';

export interface DashboardSummary {
  todaySteps: number;
  todayCalories: number;
  todayDistanceKm: number;
  todayWaterMl: number;
  weeklyWorkouts: number;
  stepGoalPercent: number;
  calorieGoalPercent: number;
  waterGoalPercent: number;
  recentWorkouts: Workout[];
  recentActivities: Activity[];
}