# lib/api-zod/src/generated/types/goalsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GoalsInput {
  stepGoal?: number;
  calorieGoal?: number;
  waterGoalMl?: number;
  workoutGoalPerWeek?: number;
}