# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  heightCm?: number | null;
  /** @nullable */
  weightKg?: number | null;
  createdAt: Date;
  updatedAt: Date;
}